Split up embed/notebook
